feature:add workspace, message and conversation schema

+++ controllers/channel.controller.js
@@ -1,7 +1,7 @@
 import { Channel } from "../models/channel.model.js"
 import { Members } from "../models/member.model.js"
+import { Workspace } from "../models/workspace.model.js"
 import logger from "../utils/logger.js"
-import axios from "axios";
 import dotenv from "dotenv";
 dotenv.config();
 
@@ -9,31 +9,33 @@ dotenv.config();
 //To create a Channel and save the details about the channel with an avatar and create a member i.e. owner as soon as the channel is created & send response.
 const createChannel = async (req, res, next) => {
     try {
-        const { name, description, userId } = req.body
-        const response = await axios.get('https://api.unsplash.com/photos/random', {
-            headers: {
-                Authorization: `Client-ID ${process.env.UNSPLASH_ACCESS_KEY}`,
-            },
-            params: {
-                count: 1,
-            },
-        });        
-        const randomImage = response.url;
+        const { workspaceId } = req.params
+        const { name, userId } = req.body
+
+        if (!workspaceId || !name || !userId){
+            return res.status(400).json({message: "All fields are required"})
+        }
+
         const newChannel = new Channel({
             name,
             members: [],
-            description,
-            avatar: randomImage,
+            workspaceId: workspaceId,
             owner: userId
         })
+
         const saveChannel = await newChannel.save()
+        
         const OwnerMember = new Members({
             userId: userId,
-            channelId: saveChannel._id
+            channelId: saveChannel._id,
+            workspaceId: workspaceId,
         })
+
         await OwnerMember.save();
+
         saveChannel.members.push(OwnerMember._id);
         await saveChannel.save();
+
         return res.status(201).json({
             message: 'Channel created successfully',
             channel: saveChannel,
@@ -50,9 +52,11 @@ const createChannel = async (req, res, next) => {
 const getAllChannels = async (req, res, next) => {
     try {
         const channels = await Channel.find().populate('members').populate('owner');
+
         if (!channels || channels.length === 0) {
             return res.status(404).json({ message: "No channels could be found" });
         }
+
         return res.status(200).json({ channels });
     } catch (error) {
         logger.error(error.message);
@@ -64,10 +68,17 @@ const getAllChannels = async (req, res, next) => {
 const getChannelUsingId = async (req, res, next) => {
     try {
         const { id } = req.params;
+
+        if (!id) {
+            return res.status(400).json({ message: "Channel ID is required" });
+        }
+
         const channel = await Channel.findById(id).populate('owner').populate('members');
+
         if (!channel) {
             return res.status(404).json({ message: "Channel not found" });
         }
+
         return res.status(200).json({ channel });
     } catch (error) {
         logger.error(error.message);
@@ -78,18 +89,24 @@ const getChannelUsingId = async (req, res, next) => {
 const updateChannel = async (req, res, next) => {
     try {
         const { id } = req.params;
-        const { name, description } = req.body;
+        const { name } = req.body;
+
+        if (!id || !name) {
+            return res.status(400).json({ message: "Channel ID and name are required" });
+        }
+
         const channel = await Channel.findById(id);
+
         if (!channel) {
             return res.status(404).json({ message: "Channel not found" });
         }
+
         if (name) {
             channel.name = name;
         }
-        if (description) {
-            channel.description = description;
-        }
+
         await channel.save();
+
         return res.status(200).json({ message: "Channel updation successful", channel });
     } catch (error) {
         logger.error(error.message);
@@ -101,11 +118,31 @@ const updateChannel = async (req, res, next) => {
 const deleteChannel = async (req, res, next) => {
     try {
         const { id } = req.params;
+
+        if (!id) {
+            return res.status(400).json({ message: "Channel ID is required" });
+        }
+
         const channel = await Channel.findByIdAndDelete(id);
+
         if (!channel) {
             return res.status(404).json({ message: "Channel was not found" });
         }
+        
         await Members.deleteMany({ channelId: id });
+
+        const {workspaceId} = channel;
+        
+        const workspace = await Workspace.findById(workspaceId);
+
+        if (!workspace) {
+            return res.status(404).json({ message: "Workspace not found" });
+        }
+
+        workspace.channels.pull(channel._id);
+
+        await workspace.save();
+
         return res.status(200).json({ message: "Channel deleted successfully" });
     } catch (error) {
         logger.error(error.message);
